End second experiment after receiving a few responses from the new server address
This allows to declare the end of the experiment in an adaptive way in the case of pre-copy and hybrid migrations, where multiple requests can be satisfied by the old server address during the pre-dump phase. With the previous method, it was mandatory to ensure that the "shutdownAfterRequest" configuration field was high enough to avoid ending the experiment in the middle of the pre-dump phase, while now the field becomes irrelevant.